% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pVal.R
\name{D_KS}
\alias{D_KS}
\title{KS statistics}
\usage{
D_KS(par, muts, L = max(muts) - min(muts) + 1, Kmin = 0)
}
\arguments{
\item{par}{vector of breakpoints}

\item{muts}{vector of mutations locations}

\item{L}{length of locus}

\item{Kmin}{minimal length of a segment}
}
\value{
sum of KS statistics
}
\description{
calculates sum of KS statistics given breakpoints.
}
\examples{
D_KS(par=c(3000,5000),muts =sort(c(sample(1:3000,400),sample(3001:8000,250),sample(8001:10000,400))),L=10000,Kmin=0)
}
