% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p0.R
\name{get_p0}
\alias{get_p0}
\title{p-value of the null model}
\usage{
get_p0(muts, L = max(muts) - min(muts) + 1, Kmin = 0, A = 20)
}
\arguments{
\item{muts}{vector of mutations locations}

\item{L}{length of locus}

\item{Kmin}{minimal length of a segment}

\item{A}{number of realizations}
}
\value{
p-value of the null model
}
\description{
estimates p-value of the null model
}
