% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBestSingleBreak}
\alias{getBestSingleBreak}
\title{function to get new best break}
\usage{
getBestSingleBreak(muts, L, parPrev)
}
\arguments{
\item{muts}{vector of mutations}

\item{L}{genome length}

\item{parPrev}{exististing breaks (including 0 and L)}
}
\value{
new break
}
\description{
function to get new best break
}
