% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreaks.R
\name{getBreaks}
\alias{getBreaks}
\title{finds optimal breaks}
\usage{
getBreaks(muts, L = max(muts) - min(muts) + 1, Kmin = 0, n = 1)
}
\arguments{
\item{muts}{vector of mutations locations}

\item{L}{length of locus}

\item{Kmin}{minimal length of a segment}

\item{n}{number of breaks}
}
\value{
results of the DE optimization
}
\description{
fits optimal breaks given number of breaks
}
