% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getChiSquare}
\alias{getChiSquare}
\title{function to get ChiSquare}
\usage{
getChiSquare(par, muts, L)
}
\arguments{
\item{par}{vector of breaks}

\item{muts}{mutations vector}

\item{L}{genome length}
}
\value{
ChiSquare
}
\description{
function to get ChiSquare
}
