% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreaks.R
\name{getNumberBreaks}
\alias{getNumberBreaks}
\title{finds optimal number of breaks}
\usage{
getNumberBreaks(
  muts,
  L = max(muts) - min(muts) + 1,
  Kmin = 0,
  pThreshold = 0.05
)
}
\arguments{
\item{muts}{vector of mutations locations}

\item{L}{length of locus}

\item{Kmin}{minimal length of a segment}

\item{pThreshold}{significance threshold of the uniform KS test}
}
\value{
breaks locations for optimal number of breaks
}
\description{
fits optimal breaks NOT given number of breaks
}
