% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreaksChiSquare.R
\name{getNumberBreaksChiSquare}
\alias{getNumberBreaksChiSquare}
\title{finds optimal number of breaks ChiSquare}
\usage{
getNumberBreaksChiSquare(muts, L = max(muts) - min(muts) + 1, Kmin = 0)
}
\arguments{
\item{muts}{vector of mutations locations}

\item{L}{length of locus}

\item{Kmin}{minimal length of a segment}
}
\value{
breaks locations for optimal number of breaks
}
\description{
fits optimal breaks NOT given number of breaks
}
