% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{improve}
\alias{improve}
\title{function to improve the breaks.}
\usage{
improve(muts, L, par)
}
\arguments{
\item{muts}{vector of mutations}

\item{L}{genome length}

\item{par}{breaks (including 0 and L)}
}
\value{
breaks
}
\description{
function to improve the breaks.
}
