% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreaksChiSquare.R
\name{getChiSquare}
\alias{getChiSquare}
\title{Chi statistics}
\usage{
getChiSquare(par, muts, L = max(muts) - min(muts) + 1, Kmin = 0, parPrev = c())
}
\arguments{
\item{par}{vector of breakpoints}

\item{muts}{vector of mutations locations}

\item{L}{length of locus}

\item{Kmin}{minimal length of a segment}
}
\value{
penalized chi square
}
\description{
calculates sum of Chi statistics given breakpoints and penalize.
}
